Colab Notebook
This repository contain a machine learning model named as KNN(k-nearest neighbors) on diabetes dataset .It is very easy ,efficient and easy to implement .suitable for both classification and regression problems. It works by identifying the "k" nearest neighbors to a new data point and assigning it to the most common class (classification) or averaging their values (regression).It is one of the popular and simplest classification and regression classifiers used in machine learning today.